clk: nexell: correct type casting
This patch corrects type casting of nexell clock driver.
It can resolve below warning:

drivers/clk/nexell/clk-s5pxx18.c:267:2: warning: format ‘%ld’ expects
argument of type ‘long int’, but argument 5 has type ‘s64’ [-Wformat=]
  pr_debug("=  %ld, %ld diff (%ld)\n", new_rate, request,

+++ drivers/clk/nexell/clk-s5pxx18.c
@@ -265,7 +265,7 @@ next:
 		 peri->name, peri->clk_step, peri->div_src_0, peri->div_src_1,
 		 freq_hz, peri->div_val_0, peri->div_val_1);
 	pr_debug("=  %ld, %ld diff (%ld)\n", new_rate, request,
-		 abs(new_rate - request));
+		 (long)abs(new_rate - request));
 
 	return new_rate;
 }
